refactor(commands): extract comment tests to sibling file
